+++ lib/screens/home_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
+import 'package:fridge/main.dart';
 import 'package:fridge/models/user.dart';
 import 'package:fridge/screens/first_page.dart';
-import 'package:fridge/screens/signin_page.dart';
 import 'package:get/get.dart';
 
 class HomePage extends StatelessWidget {
@@ -13,7 +13,7 @@ class HomePage extends StatelessWidget {
         child: Text('LOG OUT'),
         onPressed: () {
           Get.find<User>().logout();
-          Get.to(FirstPage());
+          Get.toNamed('/');
         },
       )),
     );
